event: split the interface unit
* Doxygen is dropping some types for #if #endif guard
