+++ CMakeLists.txt
@@ -53,7 +53,7 @@ add_subdirectory(modules/portable)
 if(${CMAKE_CXX_COMPILER_ID} MATCHES GNU)
     # WIP: perform test with the portable headers
 else()
-    # add_subdirectory(modules/event)     # coroutine_event
+    add_subdirectory(modules/event)     # coroutine_event
     add_subdirectory(modules/net)       # coroutine_net
     # add_subdirectory(modules/thread)    # coroutine_thread
 endif()
